refactor app considering separated domains
