+++ app.js
@@ -2,8 +2,10 @@ import express from 'express';
 import { ApolloServer } from 'apollo-server-express';
 import cors from 'cors';
 import dotenv from 'dotenv';
-import { typeDefs, resolvers } from './src/schema';
-import { getUserWithToken } from './src/util';
+import typeDefs from './src/typeDefs';
+import resolvers from './src/resolvers';
+import models from './src/models';
+// import { getUserWithToken } from './src/util';
 import './src/config';
 
 dotenv.config();
@@ -11,11 +13,14 @@ dotenv.config();
 const server = new ApolloServer({
   typeDefs,
   resolvers,
-  context: async ({ req }) => {
-    const token = req.headers.authorization || '';
-    const user = await getUserWithToken(token);
-    return { user };
+  context: {
+    models
   }
+  // context: async ({ req }) => {
+  //   const token = req.headers.authorization || '';
+  //   const user = await getUserWithToken(token);
+  //   return { user };
+  // }
 });
 
 const app = express();
